requirements.txt
PyMuPDF==1.23.4
Pillow==10.1.0
reportlab==4.0.4 